Added BGP listening port for ExtremeXOS 22.7.3.5 peers
Cisco and BIRD implementations seem fine with LB not responding on port 179
but Extreme will not re-establish after LB has shut down connection - it
goes into IDLE state. Running nc -l 179 on the LB caused it to accept
connections from LB again, so we listen and accept connection on port 179
but just leave the connection idle and close it after a minute.